[Build] Move class definition of LocalTabletsChannel to cpp file to reduce build times
