+++ be/src/runtime/lake_tablets_channel.cpp
@@ -12,7 +12,6 @@ DIAGNOSTIC_POP
 
 #include <chrono>
 #include <cstdint>
-#include <shared_mutex>
 #include <unordered_map>
 #include <vector>
